% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchaMouse16_all.R
\name{catchaMouse16_all}
\alias{catchaMouse16_all}
\title{Automatically run every time-series feature calculation included in the catchaMouse16 set.}
\usage{
catchaMouse16_all(data)
}
\arguments{
\item{data}{a numerical time-series input vector}
}
\value{
object of class DataFrame that contains the summary statistics for each feature
}
\description{
Automatically run every time-series feature calculation included in the catchaMouse16 set.
}
\author{
Trent Henderson
}
