% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FC_LoopLocalSimple_mean_stderr_chn}
\alias{FC_LoopLocalSimple_mean_stderr_chn}
\title{Function to calculate a statistical feature}
\usage{
FC_LoopLocalSimple_mean_stderr_chn(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value that denotes the calculated time-series statistic
}
\description{
Function to calculate a statistical feature
}
\examples{
x <- 1 + 0.5 * 1:1000 + arima.sim(list(ma = 0.5), n = 1000)
outs <- FC_LoopLocalSimple_mean_stderr_chn(x)

}
\author{
Trent Henderson
}
