% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_catch.R
\name{normalise_catch}
\alias{normalise_catch}
\title{Scale each value into sigmoidal range for visualisation and analysis.}
\usage{
normalise_catch(x)
}
\arguments{
\item{x}{a vector of scalar values}
}
\value{
a vector of scalar values normalised into sigmoidal range [0,1]
}
\description{
Scale each value into sigmoidal range for visualisation and analysis.
}
\examples{
data <- seq(from = 1, to = 1000, by = 1)
outs <- normalise_catch(data)

}
\author{
Trent Henderson
}
