% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PH_Walker_biasprop_05_01_sw_meanabsdiff}
\alias{PH_Walker_biasprop_05_01_sw_meanabsdiff}
\title{Function to calculate a statistical feature}
\usage{
PH_Walker_biasprop_05_01_sw_meanabsdiff(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value that denotes the calculated time-series statistic
}
\description{
Function to calculate a statistical feature
}
\examples{
x <- 1 + 0.5 * 1:1000 + arima.sim(list(ma = 0.5), n = 1000)
outs <- PH_Walker_biasprop_05_01_sw_meanabsdiff(x)

}
\author{
Trent Henderson
}
