% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_all.R
\name{catch_all}
\alias{catch_all}
\title{Automatically run every time-series feature calculation included in both catch22 and catchaMouse16 sets.}
\usage{
catch_all(data)
}
\arguments{
\item{data}{a numerical time-series input vector}
}
\value{
object of class DataFrame that contains the summary statistics for each feature
}
\description{
Automatically run every time-series feature calculation included in both catch22 and catchaMouse16 sets.
}
\author{
Trent Henderson
}
